+++ ClassLibraries/Macross.Logging.Files/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## 3.0.0
+
+* Updated for 3.0.0 version of
+  [Macross.Json.Extensions](../Macross.Json.Extensions/CHANGELOG.md) which
+  contains breaking changes.
+
 ## 2.0.0
 
 * Updated for new